requirements.txt
mariadb==1.1.14
numpy==2.3.4
packaging==25.0
pandas==2.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
tzdata==2025.2
